Subservient: v0.79 PRE-ALPHA
Subservient in a state where the single and batch video progressing fully works. At least, on Windows it's been tested over and over and should be pretty stable. In theory, it should also work for LINUX and MAC, but they still need to be thoroughly tested. This v0.79 version requires some minor bugfixing and an instructional video, and then it will go to v0.8 ALPHA phase.