v0.79 large pre-installation bugfix
"Oh no, my instructional video for Subservient is ruined because of unforeseen errors" part 2.

I believe that with these bugfixes, subservient is user friendly enough to have a relatively seamless installation process. Let's make an instructional video now, on a PC without anything related to the dependencies and packages installed.