v0.80 containerization
goodyear77 on Reddit tipped me that containerization of Subservient would make it more accessible to people, as it then becomes much easier to install. I added it to the to-do list in the ReadMe. I will work on it after I added TV-Series Support, which is scheduled to be released in the coming weeks.